Starting to add sqlite functionality. To do this I have to check we HAVE_SQLITE defined. This also means that we have to check we HAVE_MYSQL defined.
To this end I am starting to split out the specific HAVE_MYSQL parts. Expect
function names to change, new arguments to appear and old ones vanish. In
short this will be unstable. It should compile and hopefully run with
both defined, but expect lots of small bug updates if one or the other is
not.

Let the battle commence!