Hideous commit. Many changes, too many to put in here :( Basically, I have decided to create linked lists for the data in the database tables. I am also wanting to use prepared statements. You can see this in the src/cmdb_statements.h file. main() function in cmdb.c has been butchered. Very little will work with this program now. I have also decided to have checks for HAVE_LIBPCRE; if the user does not want their data validated, who am I to say they should. Will have to try at some point to add a warning in the confiugure script.
Anyway, the idea of having completely separate code for each database is
a bit tiresome. So I am gonna try and have a code base that will do
the actions and use #ifdef's for the various databases. This will probably
by ugly, but with the linked lists and prepared statements should not be
too bad, as I just pass the list to the specific database fill up
function.

More to follow