We can now add a reverse zone inot the database. The rev zonefile is written out, checked and then updated as valid in the database.
I have found a (small) but in the sqlite code, where some numbers will
overflow the int type so I have updated them to use the sqlite3_int64
type. They are then cast to unsigned long int so this should be safe
enough. This was seen with start_ip and end_ip of teh rev zones, but
as we are not using these really it was not noticed.