Modifing the build well under way.
We get the server_id then the
build_id, and then we are starting to get the id for which we want
to modify. Generally thinking that at the moment we can modify
OS, varient, partition scheme and perhaps the locale. Not looking
to be able to modify the build domain.