Preperation for updating the cmdb run_search function to use dbdata_s
The run_search function in the cmdb program does not use the dbdata_s
data type. This does not allow us to be flexible with the searches.
Currently you can only search on text and return ID's, which is fine if
we are only searching for ID's but sucks if we want some text from the
database, like a hardware class.