Adding hardware now shows the class; added cmdb_run_search
To enable a search within cmdb that was not just an id on text, I have
implemented the same style of search as in cbc, where you can specify
the number and type of fields and the number and type of arguments. This has
been done for sqlite3 and seems to work. Still need to write the mysql routines