Can modify all aspects of ldap config in cbcdomain
We can now modify the ldap authentication configuration in cbcdomain
using all combinations of basedn, binddn, ldap server and ssl.
If the user modifies the ldap config in the build_domain, then config_ldap
is automatically set to 1. This may cause issues if there is not complete
ldap authentication configuration in the DB. There is a warning message
about this; hopefully the users take heed.