More updates to zones.c to get ready to add slave zones
The list (fwd | rev) zone functions have been updated to check for
the zone type (slave or master).
The display (fwd | rev) zone functions have been updated to check for
the zone type (slave or master) and will only display the master zone
as there should be no records in the database for a slave zone.