Fixed fwd zone validation
We were checking the validity of the zones in multiple places but were
not marking the zones as valid in all of them. Therefore validate_fwd_zone
now is the only place that does validation. This calls create_and_write_fwd_zone
so this function should not be called by any other function. The validate
function will do this. Validate also marks the validity of the zone in
the database.