We can now display and write out the glue records in the zonefile.
Updated the DB code so we can select glue records now. We can also
add them to the display of the zone, and indeed write them into the
zone. It appears that you need to have the zone already configured
on the delegate server, complete with A records for nameservers so
this is something to watch out for.