Fixed bug in cbcdomain and cbc when there are no records in db.
In cbcdomain if there are no domains then we could not add one as this
was treated as an error. Similarly in cbc if there were no builds or
build_ip's this was treated as an error. This, of course, is not the
case. This is due to the cbc_run_query expecting something back from
the database and returning NO_RECORDS if there are no db records.