***Version 0.2.18.8.2*** cbcpart adds muser, mtime when adding partition.
I have added the default_part table to have the same user and time
info as the rest of the updated tables. Now, when a partition is
added to a scheme, the scheme is marked as updated in the DB and
the the partion has the user and time info added to it.
As we cannot modify a partition, muser and mtime are superflous.
This is true of some of the other DB tables.