Added function init_multi_dbdata_struct.
This is intended to be a more generic version of
init_initial_dbdata_struct because it does not rely on having access
to specific number of arguments for a database transaction. Instead
the calling function has this information and so passes the function
the number of elements in the list.