Updated check_for_disk_device to not add disk_dev to database
We now search for an already existing disk_dev and return that if
there. If not, then we build the new one, add it to details struct
and add it at the ned of create_build_config. This should stop
adding stuff to the DB if we fail with the IP address