*** Version 0.2.20 *** cbcsysp ready for business
We can now specify system build packages, organised around the build
domain, to have specific arguments to their configurations. This only
works for debian builds, and so we also need a program to specify
system configuration scripts (which are needed for both debian and
redhat based systems) with their arguments. Now, for the scripts, we
will need to specify the argument number rather than the configuration.