*** Version 0.2.22 *** cbcdomain release
cbcdomain has been updated to deal with the leaner setup. The database
has been updated to reflect the changes and we have the cbcscript and
cbcsysp programs to configure the build domain.

I have noticed a strange bug building on ubuntu 14.04 LTS where the
order of the libraries seem to make the compile fail as we are only
linked to -lpcre and not -lmysql and -lsqlite. The database libs seem
to be added too soon in the compile statement.
gcc --version
gcc (Ubuntu 4.8.2-19ubuntu1) 4.8.2